Bugfix: moved component into folder
