+++ src/component/select-option/select-option.module.scss
